// src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        "SystemStandards.Results.Result": {
            readonly isSuccess?: boolean;
            status?: components["schemas"]["SystemStandards.Results.ResultStatus"];
            readonly errors?: string[] | null;
            readonly validationErrors?: components["schemas"]["SystemStandards.Results.ValidationError"][] | null;
            readonly successMessage?: string | null;
            correlationId?: string | null;
            location?: string | null;
        };
// ...
        "SystemStandards.Results.ValidationError": {
            identifier?: string | null;
            errorMessage?: string | null;
            errorCode?: string | null;
            severity?: components["schemas"]["SystemStandards.Results.ValidationSeverity"];
        };
// ...
        "Volo.Abp.Http.RemoteServiceErrorResponse": {
            error?: components["schemas"]["Volo.Abp.Http.RemoteServiceErrorInfo"];
        };
        "Volo.Abp.Http.RemoteServiceValidationErrorInfo": {
            message?: string | null;
            members?: string[] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}